NIH Mapping

+++ src/main/Main.java
@@ -22,8 +22,8 @@ public class Main {
 		cleaner.cleanCsvFiles(inputDirectoryPath, outputDirectoryPath);
 	}
 	
-	private void mergeFiles(String intputDirectoryPath, String outputFilePath, Mode mode){
-		Merge merge = new Merge(mode);
+	private void mergeFiles(String intputDirectoryPath, String outputFilePath, String dataSource, Mode mode){
+		Merge merge = new Merge(dataSource, mode);
 		merge.mergeCsvFiles(intputDirectoryPath, outputFilePath);
 	}
 	
@@ -40,39 +40,62 @@ public class Main {
 		}			
 	}
 	
-	private static void startProcess(String url, String outputFilePath){
+//	importSource = GESIS or NIH
+	private static void startProcessWithUrl(String url, String outputFilePath, String importSource){
 		Main main = new Main();
-		String csvDirPath = outputFilePath + "_/";
-		String cleanedCsvDirPath = outputFilePath + "_cleaned/";
 		
-		System.out.println("Loading and transforming data...");
-		main.loadAndTransformData(url, csvDirPath + (new File(outputFilePath)).getName());
-		System.out.println("Done.");
-		
-		System.out.println("Preparing data for import...");
-		main.cleanData(csvDirPath, cleanedCsvDirPath);
-		System.out.println("Done.");
-		
-		if(!url.contains("EntityLink")){			
-			System.out.println("Create node csv-file for import...");
-			main.mergeFiles(cleanedCsvDirPath, outputFilePath, Mode.entity);
+		if(importSource.equals("GESIS")){
+			String csvDirPath = outputFilePath + "_/";
+			String cleanedCsvDirPath = outputFilePath + "_cleaned/";
+			
+			System.out.println("Loading and transforming data...");
+			main.loadAndTransformData(url, csvDirPath + (new File(outputFilePath)).getName());
 			System.out.println("Done.");
-		}
-		else{
-			System.out.println("Create relationship csv-file for import");
-//			main.getRelationships(csvDirPath, outputFilePath);
-			main.mergeFiles(cleanedCsvDirPath, outputFilePath, Mode.entityLink);
+			
+			System.out.println("Preparing data for import...");
+			main.cleanData(csvDirPath, cleanedCsvDirPath);
 			System.out.println("Done.");
+			
+			if(!url.contains("EntityLink")){			
+				System.out.println("Create node csv-file for import...");
+				main.mergeFiles(cleanedCsvDirPath, outputFilePath, importSource, Mode.NODES);
+				System.out.println("Done.");
+			}
+			else{
+				System.out.println("Create relationship csv-file for import");
+	//			main.getRelationships(csvDirPath, outputFilePath);
+				main.mergeFiles(cleanedCsvDirPath, outputFilePath, importSource, Mode.EDGES);
+				System.out.println("Done.");
+			}
+			System.out.println("Cleanup...");
+			main.deleteDirectory(csvDirPath);
+			main.deleteDirectory(cleanedCsvDirPath);
+			System.out.println("Finished.");
 		}
-		System.out.println("Cleanup...");
-		main.deleteDirectory(csvDirPath);
-		main.deleteDirectory(cleanedCsvDirPath);
-		System.out.println("Finished.");
+	}
+	
+	private static void startProcessWithCSV(String csvDirPath, String outputFilePath, String importSource){
+		Main main = new Main();
+
+		if(importSource.equals("NIH")){
+			main.mergeFiles(csvDirPath, outputFilePath, importSource, Mode.NODES);
+		}
+		
+	}
+	
+	private static void startProcess(String url_or_csvPath, String outputFilePath, String importSource){
+		if(url_or_csvPath.contains("http://"))
+			startProcessWithUrl(url_or_csvPath, outputFilePath, importSource);
+		else
+			startProcessWithCSV(url_or_csvPath, outputFilePath, importSource);
 	}
 	
 	public static void main(String[] args){
-		if(args != null && args.length == 2){
-			startProcess(args[0], args[1]);
-		}
+//		if(args != null && args.length == 3){
+//			startProcess(args[0], args[1], args[3]);
+//		}
+		String csvDirPath = "C:/Users/du/Projekte/LOD_Research_Graph/nih/importTest";
+		String outputFilePath = "C:/Users/du/Projekte/LOD_Research_Graph/nih/nihOutputFiles/nih.csv";
+		startProcess(csvDirPath, outputFilePath, "NIH");
 	}
 }
